ml-service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
pydantic>=2.9.0
numpy>=2.1.0
scikit-learn>=1.5.2
joblib>=1.4.2
httpx>=0.27.0
python-dotenv>=1.0.1
google-generativeai>=0.7.2
